day8/requirements.txt
fastapi
uvicorn
requests
python-dotenv
redis
